♻️ Refactor: Notification Model 패키지 변경
