+++ feature/notification/src/main/java/com/twix/notification/component/NotificationItem.kt
@@ -22,7 +22,7 @@ import com.twix.designsystem.theme.GrayColor
 import com.twix.designsystem.theme.TwixTheme
 import com.twix.domain.model.enums.AppTextStyle
 import com.twix.domain.model.enums.NotificationType
-import com.twix.domain.model.goal.Notification
+import com.twix.domain.model.notification.Notification
 import com.twix.ui.extension.noRippleClickable
 
 @Composable
