+++ feature/notification/src/main/java/com/twix/notification/contract/NotificationUiState.kt
@@ -1,6 +1,6 @@
 package com.twix.notification.contract
 
-import com.twix.domain.model.goal.Notification
+import com.twix.domain.model.notification.Notification
 import com.twix.ui.base.State
 
 data class NotificationUiState(
